reuse record comparison code from QualityIO tests
